Isolatec common code out of prjxray_form_channels.py
